registro-terminado
