+++ src/controllers/userController.js
@@ -1,6 +1,6 @@
 const path = require('path');
 const fs = require('fs');
-const { log } = require('console');
+const { log, Console } = require('console');
 const UsersFilePath= path.join(__dirname, '../data/users.json');
 
 const userController = {
@@ -15,37 +15,38 @@ const userController = {
         res.render('users/register');
       },
 
-      create:(req,res)=>{
-      let archivoUsuario = fs.readFileSync(UsersFilePath, 'utf-8');
-    function ultimoId = archivoUsuario.reduce((maxId, usuario) => {
-        return usuario.id > maxId ? usuario.id : maxId;
-      }, archivoUsuario[0].id);
-    
-    
+      create: (req, res) => {
+        let archivoUsuario = fs.readFileSync(UsersFilePath, 'utf-8');
       
-      let nuevoUsuario={
-        id: ultimoId+1,
-        nombre: req.body.usuario,
-        apellido:req.body.apellido,
-        email: req.body.email,
-        contraseña: req.body.password,
-      }
+        // Verificar si el archivo está vacío
+        if (archivoUsuario == "") {
+          archivoUsuario = [];
+        } else {
+          // Convertir la cadena de texto en un array
+          archivoUsuario = JSON.parse(archivoUsuario);
+        }
+      
+        let ultimoId = archivoUsuario.reduce((maxId, usuario) => {
+          return usuario.id > maxId ? usuario.id : maxId;
+        }, archivoUsuario[0].id);
+      
+        let nuevoUsuario = {
+          id: ultimoId + 1,
+          nombre: req.body.usuario,
+          apellido: req.body.apellido,
+          email: req.body.email,
+          contraseña: req.body.password,
+          categoria : "usuario",
+          imagen : "usuario.jpg"
 
-      let usuarios;
-      if(archivoUsuario == ""){
-         usuarios = [];
-      }
-      else{
-         usuarios = JSON.parse(archivoUsuario);
-      }
-
-      usuarios.push(nuevoUsuario);
-      usuarioJSON=JSON.stringify(usuarios);
-      fs.writeFileSync(UsersFilePath, usuarioJSON);
-
-      res.redirect("/");
-
-    }, 
+        }
+      
+        archivoUsuario.push(nuevoUsuario);
+        usuarioJSON = JSON.stringify(archivoUsuario);
+        fs.writeFileSync(UsersFilePath, usuarioJSON);
+      
+        res.redirect("/");
+      },
 
 
 
@@ -53,17 +54,23 @@ const userController = {
         let Users = JSON.parse(fs.readFileSync(UsersFilePath, 'utf-8'));
         res.render('users/userList', {Users:Users})
     },
+
+
+
     search: (req,res) => {
       const Users = JSON.parse(fs.readFileSync(UsersFilePath, 'utf-8'));
-      let searchUsers = req.query.search;
+      let searchUsers = req.query.search.toLowerCase();
       let Results = [];
-      for ( let i=1; i < Users.length;i++){
-         if(this.Users[i].name.includes(searchUsers)){
-          Results.push(this.Users[i])
-         }
+      for ( let i=0; i < Users.length;i++){
+         if(Users[i].nombre.toLowerCase().includes(searchUsers) || Users[i].apellido.toLowerCase().includes(searchUsers)){
+          Results.push(Users[i])
+        }
       }
-      return Results;
+      //res.send(Results)
+      res.render('users/userResults',{userResults : Results});
     },
+
+
     edit:(req,res)=>{
       let Users = JSON.parse(fs.readFileSync(UsersFilePath, 'utf-8'));
       let idUser= req.params.idUser;
