+++ src/data/users.json
@@ -1,12 +1,12 @@
 [
     {
         "id": 1,
-        "nombre": "Roman",
+        "nombre": "roman",
         "apellido": "Amado",
         "email": "romanamado1@gmail.com",
         "contraseña": "1231sssrr",
         "categoria": "admin",
-        "imagen": "descarga(1).jpg"
+        "imagen": "usuario.jpg"
     },
     {
         "id": 2,
@@ -15,7 +15,7 @@
         "email": "ivan02@hotmail.com",
         "contraseña": "15678922IL",
         "categoria": "usuario",
-        "imagen": "descarga(1).jpg"
+        "imagen": "usuario.jpg"
     },
     {
         "id": 3,
@@ -24,7 +24,7 @@
         "email": "huguiño2@gmail.com",
         "contraseña": "8s562123h",
         "categoria": "usuario",
-        "imagen": "descarga(1).jpg"
+        "imagen": "usuario.jpg"
     },
     {
         "id": 4,
@@ -33,7 +33,7 @@
         "email": "reynoso@hotmail.com",
         "contraseña": "33664892",
         "categoria": "usuario",
-        "imagen": "descarga(1).jpg"
+        "imagen": "usuario.jpg"
     },
     {
         "id": 5,
@@ -42,7 +42,7 @@
         "email": "gabicoronel@gmail.com",
         "contraseña": "12254762",
         "categoria": "usuario",
-        "imagen": "descarga(1).jpg"
+        "imagen": "usuario.jpg"
     },
     {
         "id": 6,
@@ -51,7 +51,7 @@
         "email": "mariangel@hotmail.com",
         "contraseña": "84422ma4",
         "categoria": "usuario",
-        "imagen": "descarga(1).jpg"
+        "imagen": "usuario.jpg"
     },
     {
         "id": 7,
@@ -60,7 +60,7 @@
         "email": "ledesma001@gmail.com",
         "contraseña": "j072019p",
         "categoria": "admin",
-        "imagen": "descarga(1).jpg"
+        "imagen": "usuario.jpg"
     },
     {
         "id": 8,
@@ -69,7 +69,7 @@
         "email": "carlosgonza@gmail.com",
         "contraseña": "16997685",
         "categoria": "usuario",
-        "imagen": "descarga(1).jpg"
+        "imagen": "usuario.jpg"
     },
     {
         "id": 9,
@@ -78,7 +78,7 @@
         "email": "jerealfo@hotmail.com",
         "contraseña": "00479251",
         "categoria": "admin",
-        "imagen": "descarga(1).jpg"
+        "imagen": "usuario.jpg"
     },
     {
         "id": 10,
@@ -87,27 +87,15 @@
         "email": "luchocapria1@gmail.com",
         "contraseña": "lc7245318",
         "categoria": "usuario",
-        "imagen": "descarga(1).jpg"
+        "imagen": "usuario.jpg"
     },
     {
-        "id": 78,
-        "nombre": "exequiel11",
-        "apellido": "juarez",
-        "email": "exequiel@digitalhouse.com",
-        "contraseña": "12345678"
-    },
-    {
-        "id": null,
-        "nombre": "usuarioprueba",
-        "apellido": "juare",
-        "email": "exe@digital.com",
-        "contraseña": "contraseña"
-    },
-    {
-        "id": null,
-        "nombre": "usuarioprueba",
-        "apellido": "Ledesma Cordoba",
-        "email": "usuarioPrueba@gmail.com",
-        "contraseña": "contraseña"
+        "id": 11,
+        "nombre": "romanji",
+        "apellido": "jimenez",
+        "email": "romanjimenez@gmail.com",
+        "contraseña": "contraseña",
+        "categoria": "usuario",
+        "imagen": "usuario.jpg"
     }
 ]
