+++ src/routes/users.routes.js
@@ -4,7 +4,7 @@ const userController = require ('../controllers/userController');
 
 router.get('/login', userController.login );
 router.get('/register', userController.registro);
-router.get('/List', userController.List);
+router.get('/admin/list', userController.List);
 router.get('/search',userController.search);
 router.get('/edit/:idUser',userController.edit);
