+++ src/views/users/userList.ejs
@@ -29,7 +29,7 @@
    <ul>
         <% for(let i=0;i<Users.length;i++) { %>
         <li>
-            <img src="<%= Users[i].imagen %>" alt="" width="200" height="150">
+            <img src="/img/users/<%= Users[i].imagen %>" alt="" width="200" height="150">
 
             <section class="UserData">
                     <p>Nombre: <%= Users[i].nombre %></p>
