+++ src/views/users/userResults.ejs
@@ -29,13 +29,13 @@
    <ul>
         <% for(let i=0;i<userResults.length;i++) { %>
         <li>
-            <img src="<%= userResults[i].imagen %>" alt="" width="200" height="150">
+            <img src="/img/users/<%= userResults[i].imagen %>" alt="" width="200" height="150">
 
             <section class="UserData">
-            <p>Nombre: <%= Users[i].nombre %></p>
-            <p>Apellido: <%= Users[i].apellido %></p>
-            <p>Correo : <%= Users[i].email %></p>
-            <p>Categoria : <%= Users[i].categoria %></p>
+            <p>Nombre: <%= userResults[i].nombre %></p>
+            <p>Apellido: <%= userResults[i].apellido %></p>
+            <p>Correo : <%= userResults[i].email %></p>
+            <p>Categoria : <%= userResults[i].categoria %></p>
             </section>
             <br>
